fix: Fetch balance on connect for Solana and Eth
